+++ backend/ezgg/src/main/java/com/matching/ezgg/domain/matchInfo/matchKeyword/dto/SupMatchParsingDto.java
@@ -8,7 +8,7 @@ import lombok.Getter;
 @Builder
 @AllArgsConstructor
 public class SupMatchParsingDto {
-	private Float visionScoreAdvantageLaneOpponent;	//0.5 이상
+	private Double visionScoreAdvantageLaneOpponent;	//0.5 이상
 	private Integer	saveAllyFromDeath; //1 이상
 	private Integer	wardPlaced;
 	private Integer	gameDuration; //wardPlaced/게임시간= 분당 1.2 이상
